requirements.txt
opencv-python
pillow
gtts
pygame
pyyaml
openai
